Show call method chooser dialog before making calls
The chooser is always shown, even if only legacy call options are
available. This is to warn people that they are about to make a call
that is not encrypted end-to-end.

If Signal or WhatsApp are available as call-options, they are presented
above legacy calls. Even if not available, Signal is always shown
either as unavailable or installable.

When making calls from searches, we don't have access to third-party
calling accounts, so the DialogFragment implementing the chooser can
asynchronously lookup the required IDs for calls to Signal and WhatsApp.

Change-Id: I641c182f256a92ef07cc835ce3a6054f2a579da6