Fix in-call RTT banner item colors
In the real-time text (RTT) screen, fix the End Call button to be red
as intended, and use neutral colors for the text and overflow menu
so that these are all actually visible in the light theme.

Switch from app:tint to android:tint; the former was not doing anything,
which may be worthy of future exploration.

Test: Manual: Make a call and tap the RTT button. The End Call button
should be visible in the banner along the top, as should the overflow
button and the call time / number dialed. They should remain visible
regardless of Dark theme or lack of Dark theme.

Issue: calyxos#2766
Change-Id: Ie531db6a0b1d30ca90d5a7bf990bc45fc17647a5